fix most of test

+++ regrid2/Lib/mvGenericRegrid.py
@@ -15,7 +15,8 @@ from distarray import MultiArrayIter
 from functools import reduce
 
 # used to locate fully masked cells
-EPS = 10*1.19209e-07
+EPS = 10 * 1.19209e-07
+
 
 def guessPeriodicity(srcBounds):
     """
@@ -31,21 +32,24 @@ def guessPeriodicity(srcBounds):
         nlon = lonsb.shape[-1]
         dlon = (lonsb.max() - lonsb.min()) / float(nlon)
         tol = 1.e-2 * dlon
-        if abs( (lonsb[..., -1] - 360.0 - lonsb[..., 0]).sum()/float(lonsb.size) ) > tol:
+        if abs((lonsb[..., -1] - 360.0 - lonsb[..., 0]
+                ).sum() / float(lonsb.size)) > tol:
             # looks like a regional model
             res = 0
     return res
+
 
 class GenericRegrid:
     """
     Generic Regrid class.
     """
+
     def __init__(self, srcGrid, dstGrid,
                  dtype,
                  regridMethod,
                  regridTool,
-                 srcGridMask = None, srcBounds = None, srcGridAreas = None,
-                 dstGridMask = None, dstBounds = None, dstGridAreas = None,
+                 srcGridMask=None, srcBounds=None, srcGridAreas=None,
+                 dstGridMask=None, dstBounds=None, dstGridAreas=None,
                  **args):
         """
         Constructor.
@@ -80,15 +84,15 @@ class GenericRegrid:
            re.search('gsreg', regridTool.lower()):
             # LibCF
             self.tool = regrid2.LibCFRegrid(srcGrid, dstGrid,
-                 srcGridMask = srcGridMask,
-                 srcBounds = srcBounds,
-                 **args)
+                                            srcGridMask=srcGridMask,
+                                            srcBounds=srcBounds,
+                                            **args)
         elif re.search('esm', regridTool.lower()):
             # ESMF
             staggerLoc = args.get('staggerLoc', 'center')
             if 'staggerLoc' in args:
                 del args['staggerLoc']
-            periodicity = args.get('periodicity', 
+            periodicity = args.get('periodicity',
                                    guessPeriodicity(srcBounds))
             if 'periodicity' in args:
                 del args['periodicity']
@@ -107,35 +111,35 @@ class GenericRegrid:
                 self.hasDstBounds = True
             self.srcGridAreasShape = None
             self.dstGridAreasShape = None
-            if srcGridAreas is not None: 
+            if srcGridAreas is not None:
                 self.srcGridAreasShape = srcGridAreas[0].shape
             if dstGridAreas is not None:
                 self.dstGridAreasShape = dstGridAreas[0].shape
 
             # Initialize
             self.tool = regrid2.ESMFRegrid(self.srcGridShape, self.dstGridShape,
-                  dtype = dtype,
-                  regridMethod = regridMethod,
-                  staggerLoc = staggerLoc,
-                  periodicity = periodicity,
-                  coordSys = coordSys,
-                  hasSrcBounds = self.hasSrcBounds,
-                  hasDstBounds = self.hasDstBounds,
-                  srcGridAreasShape= self.srcGridAreasShape,
-                  dstGridAreasShape = self.dstGridAreasShape,
-                  **args)
+                                           dtype=dtype,
+                                           regridMethod=regridMethod,
+                                           staggerLoc=staggerLoc,
+                                           periodicity=periodicity,
+                                           coordSys=coordSys,
+                                           hasSrcBounds=self.hasSrcBounds,
+                                           hasDstBounds=self.hasDstBounds,
+                                           srcGridAreasShape=self.srcGridAreasShape,
+                                           dstGridAreasShape=self.dstGridAreasShape,
+                                           **args)
             self.tool.setCoords(srcGrid, dstGrid,
-                  srcGridMask = srcGridMask,
-                  srcBounds = srcBounds,
-                  srcGridAreas = srcGridAreas,
-                  dstGridMask = dstGridMask,
-                  dstBounds = dstBounds,
-                  dstGridAreas = dstGridAreas,
-                  globalIndexing = True,
-                  **args)
+                                srcGridMask=srcGridMask,
+                                srcBounds=srcBounds,
+                                srcGridAreas=srcGridAreas,
+                                dstGridMask=dstGridMask,
+                                dstBounds=dstBounds,
+                                dstGridAreas=dstGridAreas,
+                                globalIndexing=True,
+                                **args)
         else:
             msg = """mvGenericRegrid.__init__: ERROR unrecognized tool %s,
-valid choices are: 'libcf', 'esmf'"""% regridTool
+valid choices are: 'libcf', 'esmf'""" % regridTool
             raise regrid2.RegridError(msg)
 
     def computeWeights(self, **args):
@@ -145,8 +149,8 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
         self.tool.computeWeights(**args)
 
     def apply(self, srcData, dstData,
-              rootPe = None,
-              missingValue = None,
+              rootPe=None,
+              missingValue=None,
               **args):
         """
         Regrid source to destination
@@ -159,8 +163,8 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
         """
 
         # assuming the axes are the slowly varying indices
-        srcHorizShape = srcData.shape[-self.nGridDims :]
-        dstHorizShape = dstData.shape[-self.nGridDims :]
+        srcHorizShape = srcData.shape[-self.nGridDims:]
+        dstHorizShape = dstData.shape[-self.nGridDims:]
 
         srcDataMaskFloat = None
         dstDataMaskFloat = None
@@ -177,37 +181,37 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
             # no axis... just call apply
             #
 
-
             # adjust for masking
             if missingValue is not None:
 
                 srcDataMaskFloat[:] = (srcData == missingValue)
 
-                # set field values to zero where missing, we'll add the mask 
+                # set field values to zero where missing, we'll add the mask
                 # contribution later
-                indata = numpy.array(srcData * (1 - (srcDataMaskFloat == 1)), 
+                indata = numpy.array(srcData * (1 - (srcDataMaskFloat == 1)),
                                      dtype=srcData.dtype)
 
                 # interpolate mask
-                self.tool.apply(srcDataMaskFloat, dstDataMaskFloat,  
-                                rootPe = rootPe, globalIndexing = True, **args)
+                self.tool.apply(srcDataMaskFloat, dstDataMaskFloat,
+                                rootPe=rootPe, globalIndexing=True, **args)
                 if re.search('conserv', self.regridMethod.lower(), re.I):
-                    dstMask = numpy.array((dstDataMaskFloat > 1 - EPS), numpy.int32)
+                    dstMask = numpy.array(
+                        (dstDataMaskFloat > 1 - EPS), numpy.int32)
                 else:
                     dstMask = numpy.array((dstDataMaskFloat > 0), numpy.int32)
 
                 # interpolate the data
-                self.tool.apply(indata, dstData, rootPe = rootPe, 
-                                globalIndexing = True, **args)
+                self.tool.apply(indata, dstData, rootPe=rootPe,
+                                globalIndexing=True, **args)
 
                 # add missing values
                 dstData *= (1 - dstMask)
-                dstData += dstMask*missingValue
-                
+                dstData += dstMask * missingValue
+
             else:
                 # no masking, just interpolate the data
-                self.tool.apply(srcData, dstData, rootPe = rootPe, 
-                                globalIndexing = True, **args)              
+                self.tool.apply(srcData, dstData, rootPe=rootPe,
+                                globalIndexing=True, **args)
 
         else:
 
@@ -224,7 +228,7 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
 
             # create containers to hold input/output values
             # (a copy is essential here)
-            zros = '[' + ('0,'*len(nonHorizShape)) + '...]'
+            zros = '[' + ('0,' * len(nonHorizShape)) + '...]'
             indata = numpy.array(eval('srcData' + zros))
             outdata = numpy.array(eval('dstData' + zros))
 
@@ -235,7 +239,7 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
 
                 indices = it.getIndices()
                 slce = '['
-                slce += reduce(operator.add, ['%d,'%i for i in indices])
+                slce += reduce(operator.add, ['%d,' % i for i in indices])
                 slce += '...]'
                 indata = eval('srcData' + slce)
 
@@ -244,30 +248,32 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
 
                     srcDataMaskFloat[:] = (indata == missingValue)
 
-                    # set field values to zero where missing, we'll add the mask 
+                    # set field values to zero where missing, we'll add the mask
                     # contribution later
                     indata *= (1 - (srcDataMaskFloat == 1))
 
                     # interpolate mask
                     self.tool.apply(srcDataMaskFloat, dstDataMaskFloat,
-                                    rootPe = rootPe, globalIndexing = True, **args)
+                                    rootPe=rootPe, globalIndexing=True, **args)
 
                     if re.search('conserv', self.regridMethod.lower(), re.I):
                         # cell interpolation
-                        dstMask = numpy.array((dstDataMaskFloat > 1 - EPS), numpy.int32)
+                        dstMask = numpy.array(
+                            (dstDataMaskFloat > 1 - EPS), numpy.int32)
                     else:
                         # nodal interpolation
-                        dstMask = numpy.array((dstDataMaskFloat > 0), numpy.int32)
+                        dstMask = numpy.array(
+                            (dstDataMaskFloat > 0), numpy.int32)
 
                 # interpolate the data, using the appropriate tool
-                self.tool.apply(indata, outdata, rootPe = rootPe, 
-                                globalIndexing = True, **args)
+                self.tool.apply(indata, outdata, rootPe=rootPe,
+                                globalIndexing=True, **args)
 
                 # apply missing value contribution
                 if missingValue is not None:
                     # add mask contribution
                     outdata *= (1 - dstMask)
-                    outdata += dstMask*missingValue
+                    outdata += dstMask * missingValue
 
                 # fill in dstData
                 exec('dstData' + slce + ' = outdata')
@@ -280,7 +286,7 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
         """
         return self.tool.getDstGrid()
 
-    def fillInDiagnosticData(self, diag, rootPe = None):
+    def fillInDiagnosticData(self, diag, rootPe=None):
         """
         Fill in diagnostic data
         @param diag a dictionary whose entries, if present, will be filled
@@ -288,5 +294,4 @@ valid choices are: 'libcf', 'esmf'"""% regridTool
         @param rootPe root processor where data should be gathered (or
                       None if local areas are to be returned)
         """
-        self.tool.fillInDiagnosticData(diag, rootPe = rootPe)
-
+        self.tool.fillInDiagnosticData(diag, rootPe=rootPe)
